+++ packages/alfa-array/src/array.ts
@@ -104,7 +104,7 @@ export namespace Array {
     array: ReadonlyArray<T>,
     mapper: Mapper<T, ReadonlyArray<U>, [index: number]>
   ): Array<U> {
-    const result: Array<U> = [];
+    const result = empty<U>();
 
     for (let i = 0, n = array.length; i < n; i++) {
       result.push(...mapper(array[i], i));
@@ -129,6 +129,34 @@ export namespace Array {
     return accumulator;
   }
 
+  export function reduceWhile<T, U = T>(
+    array: ReadonlyArray<T>,
+    predicate: Predicate<T, [index: number]>,
+    reducer: Reducer<T, U, [index: number]>,
+    accumulator: U
+  ): U {
+    for (let i = 0, n = array.length; i < n; i++) {
+      const value = array[i];
+
+      if (predicate(value, i)) {
+        accumulator = reducer(accumulator, value, i);
+      } else {
+        break;
+      }
+    }
+
+    return accumulator;
+  }
+
+  export function reduceUntil<T, U = T>(
+    array: ReadonlyArray<T>,
+    predicate: Predicate<T, [index: number]>,
+    reducer: Reducer<T, U, [index: number]>,
+    accumulator: U
+  ): U {
+    return reduceWhile(array, not(predicate), reducer, accumulator);
+  }
+
   export function apply<T, U>(
     iterable: ReadonlyArray<T>,
     mapper: ReadonlyArray<Mapper<T, U>>
@@ -150,7 +178,7 @@ export namespace Array {
     array: ReadonlyArray<T>,
     predicate: Predicate<T, [index: number]>
   ): Array<T> {
-    const result: Array<T> = [];
+    const result = empty<T>();
 
     for (let i = 0, n = array.length; i < n; i++) {
       const value = array[i];
@@ -238,7 +266,7 @@ export namespace Array {
     array: ReadonlyArray<T>,
     mapper: Mapper<T, Option<U>, [index: number]>
   ): Array<U> {
-    const result: Array<U> = [];
+    const result = empty<U>();
 
     for (let i = 0, n = array.length; i < n; i++) {
       for (const value of mapper(array[i], i)) {
@@ -309,7 +337,7 @@ export namespace Array {
   }
 
   export function distinct<T>(array: ReadonlyArray<T>): Array<T> {
-    const result: Array<T> = [];
+    const result = empty<T>();
 
     for (let i = 0, n = array.length; i < n; i++) {
       const value = array[i];
@@ -378,6 +406,26 @@ export namespace Array {
     ...iterables: Array<Iterable<T>>
   ): Array<T> {
     return [...Iterable.intersect(array, ...iterables)];
+  }
+
+  export function zip<T, U = T>(
+    array: ReadonlyArray<T>,
+    iterable: Iterable<U>
+  ): Array<[T, U]> {
+    const result = empty<[T, U]>();
+    const it = Iterable.iterator(iterable);
+
+    for (let i = 0, n = array.length; i < n; i++) {
+      const next = it.next();
+
+      if (next.done === true) {
+        break;
+      }
+
+      result.push([array[i], next.value]);
+    }
+
+    return result;
   }
 
   export function first<T>(array: ReadonlyArray<T>): Option<T> {